requirements.txt
# Core Kafka dependencies
kafka-python==2.0.2

# Data generation
faker==19.3.1
numpy==1.24.3

# Database
psycopg2-binary==2.9.7
sqlalchemy==2.0.20

# Data processing
pandas==2.0.3

# Visualization
plotly==5.16.1
streamlit==1.26.0

# Apache Flink (Bonus feature)
apache-flink==1.17.1

# Machine Learning (Bonus feature)
scikit-learn==1.3.0
scipy==1.11.2